requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1
Mako==1.3.9
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
cryptography==44.0.1
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.0.0
httpx==0.28.1
flake8==7.1.1
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
locust==2.32.9
python-dotenv==1.0.1
requests==2.32.3
